+++ changelog.md
@@ -11,6 +11,8 @@ Timeline:
 Widgets:
 - Fixed rowId in Table Widget.
 - Fixed disabled NumberInput slider.
+- TabSections icons can be modified using `iconClass` and `svgClass`.
+- TabSection's `onCreate` now receives also the tab content HTML Container (2nd parameter).
 
 ## 0.6.8 (master)